requirements.txt
selenium==4.9.1
pytest==7.3.1
pytest-html-4.1.1
webdriver-manager==4.0.2
allure-pytest==2.15.0